Added knife attack script
Press enter when holding the knife to enable the box collider for half a second. If the knife hits the enemy with the collider active it will deduct one point of health. Currently the way the knife is placed under the player when he is not moving is a bit wonky.